Uploaded files
if you flatpak-build these it works from commandline using `flatpak run net.veloren.airshipper` as a normal user. i tried to get the desktop shortcut / application launcher with icon file working, but it doesn't see the files when building and I can't find any good information on what exactly i have to do from here to get it to work.

the airshipper linux executable can be run on arch and fedora. it is version 0.4.2 from the releases section of SongTronix/